backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.32
openai==1.43.0
httpx==0.27.2
